+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 
 * Add Jinja template tag.
 
+* Reload when Jinja templates change.
+
 1.0.0 (2021-12-15)
 ------------------